feat(discount-validation): Enhance discount validation with status tracking and lifecycle management
- Introduce DiscountApplyResult contract for detailed validation outcomes
- Add DiscountStatus constant enum (Active, Inactive, Scheduled, Expired)
- Implement IDiscountService for centralized discount validation logic
- Enhance DiscountDto with computed Status field based on temporal constraints
- Migrate MinimumOrderAmount type from int to decimal for precision
- Improve CheckValidDiscount endpoint with comprehensive error responses
- Update discount query and repository with enhanced validation methods
- Refactor DiscountFilterModel with clarifying comments for filter purposes
- Add database migrations for discount schema refinements
- Update frontend routes and admin layout for discount management integration

+++ backend/LoLStore.API/Models/DiscountModel/DiscountFilterModel.cs
@@ -4,17 +4,24 @@ namespace LoLStore.WebAPI.Models.DiscountModel;
 
 public class DiscountFilterModel : PagingModel
 {
+    // Text search
     public string? Code { get; set; }
 
+    // Discount logic
     public decimal? DiscountValue { get; set; }
     public bool? IsPercentage { get; set; }
 
-    public int? MinimunOrderAmount { get; set; }
+    // Order constraints
+    public decimal? MinimunOrderAmount { get; set; }
+
+    // Business state (not raw DB flag)
     public bool? IsActive { get; set; }
 
+    // Validity window
     public DateTime? StartDate { get; set; }
     public DateTime? EndDate { get; set; }
 
+    // Created date filtering (admin)
     public int? Year { get; set; }
     public int? Month { get; set; }
     public int? Day { get; set; }
